+++ packages/@commitlint/commitlint-plugin-scope-ws/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/amollo-ui/amollo-lint/compare/@amollo-lint/commitlint-plugin-scope-ws@1.0.0...@amollo-lint/commitlint-plugin-scope-ws@1.1.0) (2022-03-02)
+
+**Note:** Version bump only for package @amollo-lint/commitlint-plugin-scope-ws
+
+
+
+
+
 # 1.0.0 (2022-02-26)