+++ packages/@eslint/eslint-config-jsx/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/amollo-ui/amollo-lint/compare/@amollo-lint/eslint-config-jsx@1.0.0...@amollo-lint/eslint-config-jsx@1.1.0) (2022-03-01)
+
+
+### Features
+
+* **@amollo-lint:eslint-jsx:** remove overrides ([390a02f](https://github.com/amollo-ui/amollo-lint/commit/390a02f9b8267943ffd91fcdcf6ca7a33f3d2a79))
+
+
+
+
+
 # 1.0.0 (2022-02-28)